src/main/java/com/yrj/service/impl/CategoryServiceImpl.java
package com.yrj.service.impl;

import com.yrj.dao.ProductCategoryRepository;
import com.yrj.entity.ProductCategory;
import com.yrj.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer id) {
        return repository.getOne(id);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryType(List<Integer> list) {
        return repository.findProductCategoriesByCategoryTypeIn(list);
    }

    @Override
    public ProductCategory save(ProductCategory category) {
        return repository.save(category);
    }
}
